[FEM] Elmer writer: sort out elasticity equation
- sort out elasticity equation - last step to refactor writer.py
