+++ src/Mod/Fem/CMakeLists.txt
@@ -252,6 +252,7 @@ SET(FemSolverElmer_SRCS
 SET(FemSolverElmerEquations_SRCS
     femsolver/elmer/equations/__init__.py
     femsolver/elmer/equations/elasticity.py
+    femsolver/elmer/equations/elasticity_writer.py
     femsolver/elmer/equations/electricforce.py
     femsolver/elmer/equations/electricforce_writer.py
     femsolver/elmer/equations/electrostatic.py
